Active Navigation Menu On Page Scroll Issue
The active menu item on scroll logic has an issue or bug if you change the css from position fixed to position sticky and also if you modify the margin property in relation to the body element. I've encountered this before and would have to search through he repository of this feature or my YouTube videos regarding it. The problem this causes is that if you select an active menu item and it scrolls to it, it won't bring you to the very top and it will be somewhat chopped off. To compensate for this, I dynamically changed the height property of the container so that it isn't chopped off at the top. The next issue this would cause is that there would be too much space at the top if you scrolled up further and affecting the layout or spacing. The solution to that is that if you do in fact scroll up, we will dynamically change the height from the previously added value of height:110vh to height: auto and it removes the excess space.