Upload Image Feature - Front End
I instructed ChatGPT to add the ability to preview an image before uploading. This is in response to a previous instruction/prompt which is as follows: "modify the following code to facilitate the uploading of an image:". ChatGPT's explanation is as follows:
The modifications made to the HTML form are as follows:

Added enctype="multipart/form-data" to the <form> tag. This attribute is necessary for uploading files.

Added a new <input> element with type="file" and name="image". This allows users to select an image file for uploading.

Added an <img> element with id="imagePreview" and style="display: none;". This element will be used to display a preview of the selected image.

Added an onchange event to the <input> element for the image. This event calls the previewImage() function when the user selects an image file.

Added a <script> section that defines the previewImage() function. This function retrieves the selected image file and sets the src attribute of the imagePreview element to the URL of the selected image, thus displaying a preview of the image.

Please note that the image preview functionality is implemented using JavaScript, and the preview will be displayed only after the user selects an image file.