Image Updates
It includes making the form capable of sending files, previewing an image after it has been selected and a placeholder image. The placeholder image was created through https://placehold.co/ and the logic itself was created through ChatGPT. The instruction/prompt was the following: create a ternary operator with the src attribute with the following code:    <img id="imagePreview" src="<?php echo './assets/images/'.$row['image'] ?>" alt="Image Preview">. The response was the following: In this code, the ternary operator is used to conditionally set the src attribute of the <img> tag. If $row['image'] is not empty or false, it will be used as the source of the image. Otherwise, it will fall back to a placeholder image (in this example, the placeholder image is named "placeholder.jpg"). Adjust the placeholder image filename accordingly to match the actual filename of your placeholder image.