All Tours Section Implemented
Both the front and back end of this section have been implemented but it's not entirely finished. The missing aspect is showing a discount if one was entered in the backend and a booking button. The CSS has to be adjusted to give uniformity to the part of each card that has the text and the button.