Refer To The Previous Commit
I removed everything from index.php and there was no code/html available when inspecting it via the Google Developer Console which suggests the auto deployment is in fact working and updating the website based off what comes from the Github repository/commit. So the problem may be with what is being uploaded in the file itself, perhaps an error in the code? I'll comment out the most recent thing added and if the page loads as once before it then means that the problem is with what is being uploaded.